requirements.txt
gTTS==2.5.1
llamaapi==0.1.36
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
Pillow==10.3.0
pygame==2.5.2
python-dotenv==1.0.1
Requests==2.32.2
